Improve Flow coverage for src files
