+++ flow-typed/globals.js
@@ -1,4 +1,10 @@
-// @flow
+/** Copyright (c) 2018 Uber Technologies, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
 
 declare var __NODE__: boolean;
 declare var __BROWSER__: boolean;
