+++ src/jwt-browser.js
@@ -2,6 +2,13 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
  */
 
-export default null;
+import type {FusionPlugin} from 'fusion-core';
+import type {Session} from 'fusion-tokens';
+
+import type {SessionDeps} from './types.js';
+
+export default ((null: any): FusionPlugin<SessionDeps, Session>);
